LoopFrom statement now takes a variable expression instead of an entire expression Fixed an issue in the scanner concerning float literals Added type checking for binary expressions, set expressions, and almost all statements. Type checker is not complete

+++ Interpreter/Scanner/Scanner.swift
@@ -252,7 +252,7 @@ class Scanner {
         let literalString = source[start..<current]
         
         if isDouble {
-            addToken(type: .DOUBLE, value: Double(literalString))
+            addToken(type: .FLOAT, value: Double(literalString))
         } else {
             addToken(type: .INTEGER, value: Int(literalString))
         }
